+++ packaging/asdf-oratos/bin/lib.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+# asdf plugin metadata
+# shellcheck disable=SC2034
+ASDF_ORATOS_REPO="latentmeta/oratos"